UI and UX rework for layers management and device association
