+++ src/IoTHub.Portal.Client/Dialogs/Layers/DeleteLayerDialog.razor
