--- src/IoTHub.Portal.Client/Pages/Layer/LayerListPage.razor
@@ -1,176 +1,0 @@
-﻿@page "/layer"
-@inherits AuthorizedComponentBase
-
-@attribute [Authorize]
-
-@inject ILayerClientService LayerClientService
-@inject IDeviceClientService DeviceClientService
-@inject IDialogService DialogService
-
-<AuthorizedContent IsLoading="@IsLoading" IsAuthorized="@IsAuthorized">
-    <MudGrid>
-        <MudItem xs="12">
-            <MudExpansionPanel Text="Layer" IsInitiallyExpanded="true">
-                <TitleContent><MudText Typo="Typo.h6">Layer</MudText></TitleContent>
-                <ChildContent>
-                    <MudTreeView Items="@Layers" Hover="true" Width="100%" ExpandOnClick="true">
-                        <ItemTemplate Context="item">
-                            <MudTreeViewItem Items="@item.Children" @bind-Expanded="@item.IsExpanded">
-                                <BodyContent>
-                                    <div id="editLayerElement" style="display: grid; grid-template-columns: 1fr auto; align-items: center; width: 100%" @onmouseover="() => PrintDetails = item.LayerData.Id">
-                                        <MudText>@item.LayerData.Name</MudText>
-                                        <div style="justify-self: end;">
-                                            @if (PrintDetails == item.LayerData.Id)
-                                            {
-                                                <MudIconButton id="editLayerDeleteLayer" Icon="@(item.LayerData.Father != null ? Icons.Material.Filled.Delete : @Icons.Material.Outlined.Delete)" Size="Size.Medium" Color="Color.Inherit" @onclick="() => DeleteLayer(item, Layers)" />
-                                                <MudIconButton id="editLayerAddLayer" Icon="@Icons.Material.Filled.AddBox" Size="Size.Medium" Color="Color.Inherit" @onclick="() => AddLayer(item)" />
-                                            }
-                                            <MudIconButton id="editLayerDevice" Icon="@Icons.Material.Filled.Edit" Size="Size.Medium" Color="Color.Inherit" @onclick="() => OpenDevices(item.LayerData)" />
-                                        </div>
-                                    </div>
-                                </BodyContent>
-                            </MudTreeViewItem>
-                        </ItemTemplate>
-                    </MudTreeView>
-                </ChildContent>
-            </MudExpansionPanel>
-        </MudItem>
-    </MudGrid>
-</AuthorizedContent>
-
-@code {
-    protected override PortalPermissions[] RequiredPermissions { get; } = { PortalPermissions.LayerRead };
-
-    [CascadingParameter]
-    public Error Error { get; set; } = default!;
-
-    public HashSet<LayerHash> Layers { get; set; } = new();
-
-    private IEnumerable<DeviceListItem> Devices { get; set; } = new List<DeviceListItem>();
-
-    public string? PrintDetails { get; set; }
-
-    protected override async Task OnInitializedAsync()
-    {
-        await base.OnInitializedAsync();
-
-        var listLayers = await LayerClientService.GetLayers();
-        var result = await DeviceClientService.GetDevices("api/devices?pageNumber=0&pageSize=10000");
-        Devices = result.Items;
-
-        if (listLayers.Count == 0)
-        {
-            try
-            {
-                LayerDto newLayer = new LayerDto
-                {
-                    Name = "Main Layer",
-                    Father = null,
-                    Planning = "None"
-                };
-                newLayer.Id = await LayerClientService.CreateLayer(newLayer);
-                Layers.Add(new LayerHash(newLayer, 0, false));
-            }
-            catch (ProblemDetailsException exception)
-            {
-                Error.ProcessProblemDetails(exception);
-            }
-        }
-        Layers = LayerHelper.GetHashsetLayer(listLayers);
-    }
-
-    public async Task AddLayer(LayerHash layer)
-    {
-        try
-        {
-            LayerDto newLayer = new LayerDto
-            {
-                Name = "New Layer",
-                Father = layer.LayerData.Id,
-                Planning = layer.LayerData.Planning != null ? layer.LayerData.Planning : "None"
-            };
-            newLayer.Id = await LayerClientService.CreateLayer(newLayer);
-
-            layer.Children.Add(new LayerHash(newLayer, layer.Level + 1, false));
-            layer.IsExpanded = true;
-        }
-        catch (ProblemDetailsException exception)
-        {
-            Error.ProcessProblemDetails(exception);
-        }
-    }
-
-    public async Task DeleteLayer(LayerHash delLayer, HashSet<LayerHash> fatherLayers)
-    {
-        foreach (var layer in fatherLayers)
-        {
-            if (layer.Children.Contains(delLayer))
-            {
-                await RemoveChild(delLayer);
-                layer.Children.Remove(delLayer);
-                StateHasChanged();
-            }
-            else
-            {
-                await DeleteLayer(delLayer, layer.Children);
-            }
-        }
-    }
-
-    public async Task RemoveChild(LayerHash delLayer)
-    {
-        if (delLayer.Children.Count() != 0)
-            foreach (LayerHash layer in delLayer.Children) await RemoveChild(layer);
-        try
-        {
-            await DeleteLayerOnDevice(delLayer.LayerData.Id);
-            await LayerClientService.DeleteLayer(delLayer.LayerData.Id);
-        }
-        catch (ProblemDetailsException exception)
-        {
-            Error.ProcessProblemDetails(exception);
-        }
-    }
-
-    public async Task DeleteLayerOnDevice(string delLayerId)
-    {
-        foreach (var device in Devices.Where(device => device.LayerId == delLayerId))
-        {
-            var updatedDevice = FindDevice(device.DeviceID);
-            updatedDevice.LayerId = null;
-            await DeviceClientService.UpdateDevice(updatedDevice);
-        }
-    }
-
-    private DeviceDetails FindDevice(string deviceId)
-    {
-        var device = Devices.FirstOrDefault(device => device.DeviceID == deviceId);
-
-        var deviceDetails = new DeviceDetails();
-
-        if (device == null) return deviceDetails;
-
-        deviceDetails.DeviceID = device.DeviceID;
-        deviceDetails.DeviceName = device.DeviceName;
-        deviceDetails.ModelId = device.DeviceModelId;
-        deviceDetails.Image = device.Image;
-        deviceDetails.IsEnabled = device.IsEnabled;
-        deviceDetails.LastActivityTime = device.LastActivityTime;
-        deviceDetails.Labels = device.Labels.ToList();
-        deviceDetails.LayerId = device.LayerId;
-
-        return deviceDetails;
-    }
-
-    public async void OpenDevices(LayerDto layer)
-    {
-        var parameter = new DialogParameters { { "InitLayer", layer }, { "LayerList", Layers } };
-
-        DialogOptions options = new DialogOptions() { MaxWidth = MaxWidth.Medium, FullWidth = true, CloseOnEscapeKey = true, CloseButton = true };
-
-        _ = await DialogService.Show<LinkDeviceLayerDialog>("Select your " + layer.Name + " devices", parameter, options).Result;
-
-        var newDeviceList = await DeviceClientService.GetDevices("api/devices?pageNumber=0&pageSize=10000");
-        Devices = newDeviceList.Items;
-    }
-}
